# doc/api.yaml
openapi: 3.0.0
paths:
    /: {get: {operationId: AppController_getHello, parameters: [], responses: {'200': {description: ""}}, tags: [App]}}
    /user: {post: {operationId: UserController_create, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateUserDto'}}}}, responses: {'201': {description: ""}}, tags: [user]}, get: {operationId: UserController_findAll, parameters: [], responses: {'200': {description: ""}}, tags: [user]}}
    '/user/{id}': {get: {operationId: UserController_findOne, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'200': {description: ""}}, tags: [user]}, put: {operationId: UserController_update, parameters: [{name: id, required: true, in: path, schema: {type: string}}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdatePasswordDto'}}}}, responses: {'200': {description: ""}}, tags: [user]}, delete: {operationId: UserController_remove, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'204': {description: ""}}, tags: [user]}}
    /track: {post: {operationId: TracksController_create, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateTrackDto'}}}}, responses: {'201': {description: ""}}, tags: [track]}, get: {operationId: TracksController_findAll, parameters: [], responses: {'200': {description: ""}}, tags: [track]}}
    '/track/{id}': {get: {operationId: TracksController_findOne, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'200': {description: ""}}, tags: [track]}, put: {operationId: TracksController_update, parameters: [{name: id, required: true, in: path, schema: {type: string}}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateTrackDto'}}}}, responses: {'200': {description: ""}}, tags: [track]}, delete: {operationId: TracksController_remove, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'204': {description: ""}}, tags: [track]}}
    /artist: {post: {operationId: ArtistsController_create, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateArtistDto'}}}}, responses: {'201': {description: ""}}, tags: [artist]}, get: {operationId: ArtistsController_findAll, parameters: [], responses: {'200': {description: ""}}, tags: [artist]}}
    '/artist/{id}': {get: {operationId: ArtistsController_findOne, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'200': {description: ""}}, tags: [artist]}, put: {operationId: ArtistsController_update, parameters: [{name: id, required: true, in: path, schema: {type: string}}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateArtistDto'}}}}, responses: {'200': {description: ""}}, tags: [artist]}, delete: {operationId: ArtistsController_remove, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'204': {description: ""}}, tags: [artist]}}
    /album: {post: {operationId: AlbumsController_create, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateAlbumDto'}}}}, responses: {'201': {description: ""}}, tags: [album]}, get: {operationId: AlbumsController_findAll, parameters: [], responses: {'200': {description: ""}}, tags: [album]}}
    '/album/{id}': {get: {operationId: AlbumsController_findOne, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'200': {description: ""}}, tags: [album]}, put: {operationId: AlbumsController_update, parameters: [{name: id, required: true, in: path, schema: {type: string}}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateAlbumDto'}}}}, responses: {'200': {description: ""}}, tags: [album]}, delete: {operationId: AlbumsController_remove, parameters: [{name: id, required: true, in: path, schema: {type: string}}], responses: {'204': {description: ""}}, tags: [album]}}
    /favs: {get: {operationId: FavoritesController_findAll, parameters: [], responses: {'200': {description: ""}}, tags: [favs]}}
    '/favs/{entity}/{id}': {post: {operationId: FavoritesController_createFavorite, parameters: [{name: entity, required: true, in: path, schema: {type: string}}, {name: id, required: true, in: path, schema: {type: string}}], responses: {'201': {description: ""}}, tags: [favs]}, delete: {operationId: FavoritesController_removeFavorite, parameters: [{name: entity, required: true, in: path, schema: {type: string}}, {name: id, required: true, in: path, schema: {type: string}}], responses: {'204': {description: ""}}, tags: [favs]}}
    /auth/login: {post: {operationId: AuthController_login, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginDto'}}}}, responses: {'201': {description: ""}}, tags: [Auth]}}
    /auth/signup: {post: {operationId: AuthController_signup, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateUserDto'}}}}, responses: {'201': {description: ""}}, tags: [Auth]}}
    /auth/refresh: {post: {operationId: AuthController_refresh, parameters: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RefreshTokenDto'}}}}, responses: {'201': {description: ""}}, tags: [Auth]}}
info:
    title: 'API Documentation'
    description: 'Auto-generated API specification'
    version: '1.0'
    contact: {}
tags: []
servers: []
components:
    securitySchemes: {bearer: {scheme: bearer, bearerFormat: JWT, type: http}}
    schemas: {CreateUserDto: {type: object, properties: {}}, UpdatePasswordDto: {type: object, properties: {}}, CreateTrackDto: {type: object, properties: {name: {type: string, example: 'Billie Jean', description: 'Track name'}, duration: {type: number, example: 300, description: 'Track duration in seconds'}, artistId: {type: string, example: a1b2c3d4-e5f6-7890-abcdef123456, description: 'Artist ID'}, albumId: {type: string, example: b7c8d9e0-f1g2-3456-hijk78901234, description: 'Album ID'}}, required: [name, duration]}, UpdateTrackDto: {type: object, properties: {name: {type: string, example: 'Billie Jean', description: 'Track name'}, duration: {type: number, example: 300, description: 'Track duration in seconds'}, artistId: {type: string, example: a1b2c3d4-e5f6-7890-abcdef123456, description: 'Artist ID'}, albumId: {type: string, example: b7c8d9e0-f1g2-3456-hijk78901234, description: 'Album ID'}}}, CreateArtistDto: {type: object, properties: {name: {type: string, example: 'Michael Jackson', description: 'Artist name'}, grammy: {type: boolean, example: true, description: 'Has Grammy award'}}, required: [name, grammy]}, UpdateArtistDto: {type: object, properties: {name: {type: string, example: 'Michael Jackson', description: 'Artist name'}, grammy: {type: boolean, example: true, description: 'Has Grammy award'}}}, CreateAlbumDto: {type: object, properties: {name: {type: string, example: Thriller, description: 'Album name'}, year: {type: number, example: 1982, description: 'Release year'}, artistId: {type: string, example: a1b2c3d4-e5f6-7890-abcdef123456, description: 'Artist ID'}}, required: [name, year]}, UpdateAlbumDto: {type: object, properties: {name: {type: string, example: Thriller, description: 'Album name'}, year: {type: number, example: 1982, description: 'Release year'}, artistId: {type: string, example: a1b2c3d4-e5f6-7890-abcdef123456, description: 'Artist ID'}}}, LoginDto: {type: object, properties: {}}, RefreshTokenDto: {type: object, properties: {}}}